+++ R/jobs.R
@@ -80,7 +80,7 @@ bqr_wait_for_job <- function(job, wait=5){
 #' Poll a jobId
 #' 
 #' @param projectId projectId of job
-#' @param jobId jobId to poll
+#' @param jobId jobId to poll, or a job Object
 #' 
 #' @return A Jobs resource
 #' 
@@ -134,8 +134,12 @@ bqr_wait_for_job <- function(job, wait=5){
 #' 
 #' @family BigQuery asynch query functions  
 #' @export
-bqr_get_job <- function(projectId = bqr_get_global_project(), jobId){
+bqr_get_job <- function(jobId, projectId = bqr_get_global_project()){
   check_bq_auth()
+  
+  if(is.job(jobId)){
+    jobId <- jobId$jobReference$jobId
+  }
   stopifnot(inherits(projectId, "character"),
             inherits(jobId, "character"))